--- src/main/java/org/sopt36/ninedotserver/auth/repository/AuthProviderRepositoryCustom.java
@@ -1,5 +1,0 @@
-package org.sopt36.ninedotserver.auth.repository;
-
-public interface AuthProviderRepositoryCustom {
-
-}
