--- src/main/java/org/sopt36/ninedotserver/auth/repository/RefreshTokenRepositoryCustom.java
@@ -1,5 +1,0 @@
-package org.sopt36.ninedotserver.auth.repository;
-
-public interface RefreshTokenRepositoryCustom {
-
-}
